% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{calibration_parameters}
\alias{calibration_parameters}
\title{Parameters needed to calibrate the cluster tree}
\usage{
calibration_parameters()
}
\description{
Parameters needed to calibrate the cluster tree
}
\keyword{internal}
