% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{cluster_graph}
\alias{cluster_graph}
\title{The beliefs for the clusters and sepsets of a cluster tree,
of mixed continuous types.}
\usage{
cluster_graph(beliefs)
}
\arguments{
\item{beliefs}{the vector of continuous beliefs to put on the cluster tree}
}
\description{
The beliefs for the clusters and sepsets of a cluster tree,
of mixed continuous types.
}
\keyword{internal}
