% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmFit.R
\name{coef.glmmFit}
\alias{coef.glmmFit}
\title{Extract Model Coefficients for a glmmFit object}
\usage{
\method{coef}{glmmFit}(object, ...)
}
\arguments{
\item{object}{glmmFit object}

\item{...}{ignored}
}
\value{
A names vector of model coefficients
}
\description{
Extract Model Coefficients for a glmmFit object
}
\keyword{internal}
