% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{continuous_beliefs}
\alias{continuous_beliefs}
\title{A vector of terms in the factorization of a graphical model,
of mixed continuous types.}
\usage{
continuous_beliefs()
}
\description{
A vector of terms in the factorization of a graphical model,
of mixed continuous types.
}
\keyword{internal}
