% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmFit.R
\name{find_approximation_name}
\alias{find_approximation_name}
\title{Find the name of the likelihood approximation used for fitting}
\usage{
find_approximation_name(method, control)
}
\arguments{
\item{method}{the method used to approximate the likelihood. The options
are \code{"Laplace"}, \code{"AGQ"} (the adaptive Gaussian quadrature approximation,
from \code{lme4}), \code{"SR"} (the sequential reduction approximation)
and \code{"IS"} (an importance sampling approximation).}

\item{control}{a list of extra parameters controlling the approximation
to the likelihood. See 'Details' for more information.}
}
\value{
a string with the name of the likelihood approximation
}
\description{
Find the name of the likelihood approximation used for fitting
}
\keyword{internal}
