% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_devfun.R
\name{find_lfun_glmm}
\alias{find_lfun_glmm}
\title{Find the log-likelihood function}
\usage{
find_lfun_glmm(modfr, method, control)
}
\arguments{
\item{modfr}{a model frame, the output of \code{find_modfr_glmm}}

\item{control}{a list of extra parameters controlling the approximation
to the likelihood}
}
\description{
Find the log-likelihood function
}

