% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmerSRMod.R
\name{glmerSRMod}
\alias{glmerSRMod}
\title{Construct a glmerSRMod object}
\usage{
glmerSRMod(x)
}
\arguments{
\item{x}{a list}
}
\value{
An object of class \code{glmerSRMod}
}
\description{
Construct a glmerSRMod object
}

