% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{glmm}
\alias{glmm}
\title{Fit a GLMM}
\usage{
glmm(formula, subformula = NULL, data = NULL, family = gaussian, method,
  control = list(), weights = NULL, offset = NULL, devFunOnly = FALSE)
}
\arguments{
\item{formula}{a two-sided linear formula object describing both the
    fixed-effects and fixed-effects part of the model, with the response
    on the left of a \code{~} operator and the terms, separated by
    \code{+} operators, on the right.  Random-effects terms are
    distinguished by vertical bars (\code{"|"}) separating expressions
    for design matrices from grouping factors.}

\item{subformula}{a subformula, describing how a substituted variable
depends on covariates, or a list of subformulas, if there is more
than one \code{Sub()} term in \code{formula}.}

\item{data}{an optional data frame, list or environment containing the
variables named in \code{formula}, and in any of the subformulas.}

\item{family}{a GLM family, see \code{\link[stats]{glm}} and
    \code{\link[stats]{family}}.}

\item{control}{a list of extra parameters controlling the approximation
to the likelihood}

\item{weights}{an optional vector of \sQuote{prior weights} to be used
      in the fitting process.  Should be \code{NULL} or a numeric
      vector.}

\item{offset}{this can be used to specify an \emph{a priori} known
    component to be included in the linear predictor during
    fitting. This should be \code{NULL} or a numeric vector of length
    equal to the number of cases.  One or more \code{\link{offset}}
    terms can be included in the formula instead or as well, and if more
    than one is specified their sum is used.  See \code{\link{model.offset}}.}

\item{devFunOnly}{logical - return only the deviance evaluation
    function. Note that because the deviance function operates on
    variables stored in its environment, it may not return
    \emph{exactly} the same values on subsequent calls (but the results
    should always be within machine tolerance).}
}
\description{
Fit a GLMM
}

