% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmFit.R
\name{glmmFit}
\alias{glmmFit}
\title{Construct a glmmFit object}
\usage{
glmmFit(x)
}
\arguments{
\item{x}{a list}
}
\value{
An object of class \code{glmmFit}
}
\description{
Construct a glmmFit object
}
\keyword{internal}
