% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/glmmsr.R
\docType{package}
\name{glmmsr}
\alias{glmmsr}
\alias{glmmsr-package}
\title{glmmsr: fit GLMMs with the sequential reduction approximation}
\description{
glmmsr allows fitting of a generalized linear mixed model, using the
sequential reduction approximation to the likelihood.
}
\details{
The interface and output are designed to look similar to lme4.

The lme4 package uses the Laplace approximation to the likelihood, except in
the special case of two-level models, where adapative Gaussian quadrature may
be used. In some situations, this Laplace approximation may be poor quality,
and the sequential reduction approximation may be used to find an accurate
approximation to the likelihood.

The main function of the glmmsr package is \code{\link{glmerSR}}, which is
used to fit the GLMM. Its interface is a generalization of
\code{\link[lme4]{glmer}}, to allow a larger class of models, including
structured pairwise comparison models. This interface is described in
\code{\link{subform}}.
}
\references{
Helen E. Ogden (2015). A sequential reduction method for
  inference in generalized linear mixed models. Electronic Journal of
  Statistics 9: 135-152. doi:
  \href{http://dx.doi.org/10.1214/15-EJS991}{10.1214/15-EJS991}
}

