% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_devfun.R
\name{mkGlmerDevfunSR}
\alias{mkGlmerDevfunSR}
\title{Make the deviance function}
\usage{
mkGlmerDevfunSR(fr, X, reTrms, family, nAGQ = 1L, k = 0L, verbose = 0L,
  control = glmmsrControl(), ...)
}
\arguments{
\item{family}{a GLM family, see \code{\link[stats]{glm}} and
    \code{\link[stats]{family}}.}

\item{nAGQ}{integer scalar - the number of points per axis for
    evaluating the adaptive Gauss-Hermite approximation to the
    log-likelihood.  Defaults to 1, corresponding to the Laplace
    approximation.  Values greater than 1 produce greater accuracy in
    the evaluation of the log-likelihood at the expense of speed.  A
    value of zero uses a faster but less exact form of parameter
    estimation for GLMMs by optimizing the random effects and the
    fixed-effects coefficients in the penalized iteratively reweighted
    least squares step.}

\item{k}{integer scalar - the level of approximation used in the sequential
reduction approximation to the likelihood.}

\item{verbose}{integer scalar.  If \code{> 0} verbose output is
    generated during the optimization of the parameter estimates.  If
    \code{> 1} verbose output is generated during the individual PIRLS
    steps.}

\item{control}{the output of a call to \code{\link{glmmsrControl}}, a list
containing control paramaters.}
}
\description{
Make the deviance function
}

