% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimization.R
\name{optimizeGlmerSR}
\alias{optimizeGlmerSR}
\title{Maximize the approximated log-likelihood}
\usage{
optimizeGlmerSR(devfun, p_beta, p_theta, init_na = NULL, verbose = 0L)
}
\arguments{
\item{devfun}{-2*loglikelihood}

\item{p_beta}{the number of covariates}

\item{p_theta}{the number of random effects}

\item{init_na}{a normal approximation to the log-likelihood surface}

\item{verbose}{integer. If verbose=0, nothing is printed. If verbose=1, a
dot is printed for each likelihood evaluation. If verbose=2, the values of
the parameters and the likelihood are printed for each evaluation.}
}
\value{
A list, containing the parameter estimate and variance matrix
}
\description{
Find the approximate maximum likelihood estimate, and finds
a normal approximation to the likelihood surface
}

