% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmerSRMod.R
\name{print.glmerSRMod}
\alias{print.glmerSRMod}
\title{Print glmerSRMod object}
\usage{
\method{print}{glmerSRMod}(x, ...)
}
\arguments{
\item{x}{glmerSRMod object}

\item{...}{ignored}
}
\description{
Print glmerSRMod object
}

