% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmFit.R
\name{print.glmmFit}
\alias{print.glmmFit}
\title{Print glmmFit object}
\usage{
\method{print}{glmmFit}(x, ...)
}
\arguments{
\item{x}{glmmFit object}

\item{...}{ignored}
}
\description{
Print glmmFit object
}
\keyword{internal}
