% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmMod.R
\name{print.glmmMod}
\alias{print.glmmMod}
\title{Print glmmMod object}
\usage{
\method{print}{glmmMod}(x, ...)
}
\arguments{
\item{x}{glmmMod object}

\item{...}{ignored}
}
\description{
Print glmmMod object
}

