% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmerSRMod.R
\name{print.summaryGlmerSRMod}
\alias{print.summaryGlmerSRMod}
\title{Print summaryGlmerSRMod object}
\usage{
\method{print}{summaryGlmerSRMod}(x, ...)
}
\arguments{
\item{x}{summaryGlmerSRMod object}

\item{...}{ignored}
}
\description{
Print summaryGlmerSRMod object
}

