% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmFit.R
\name{print.summaryGlmmFit}
\alias{print.summaryGlmmFit}
\title{Print summaryGlmmFit object}
\usage{
\method{print}{summaryGlmmFit}(x, ...)
}
\arguments{
\item{x}{summaryGlmmFit object}

\item{...}{ignored}
}
\description{
Print summaryGlmmFit object
}
\keyword{internal}
