% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmMod.R
\name{print.summaryGlmmMod}
\alias{print.summaryGlmmMod}
\title{Print summaryGlmmMod object}
\usage{
\method{print}{summaryGlmmMod}(x, ...)
}
\arguments{
\item{x}{summaryGlmmMod object}

\item{...}{ignored}
}
\description{
Print summaryGlmmMod object
}

