% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmerSRMod.R
\name{summary.glmerSRMod}
\alias{summary.glmerSRMod}
\title{Summarize a glmerSRMod fit}
\usage{
\method{summary}{glmerSRMod}(object, ...)
}
\arguments{
\item{object}{glmerSRMod object}

\item{...}{ignored}
}
\value{
An object of class \code{summaryGlmerSRMod}
}
\description{
Summarize a glmerSRMod fit
}

