% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmFit.R
\name{summary.glmmFit}
\alias{summary.glmmFit}
\title{Summarize a glmmFit object}
\usage{
\method{summary}{glmmFit}(object, ...)
}
\arguments{
\item{object}{glmmFit object}

\item{...}{ignored}
}
\value{
An object of class \code{summaryGlmmFit}
}
\description{
Summarize a glmmFit object
}
\keyword{internal}
