% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmMod.R
\name{summary.glmmMod}
\alias{summary.glmmMod}
\title{Summarize a glmmMod fit}
\usage{
\method{summary}{glmmMod}(object, ...)
}
\arguments{
\item{object}{glmmMod object}

\item{...}{ignored}
}
\value{
An object of class \code{summaryGlmmMod}
}
\description{
Summarize a glmmMod fit
}

