% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmerSRMod.R
\name{summaryGlmerSRMod}
\alias{summaryGlmerSRMod}
\title{Construct a summaryGlmerSRMod object}
\usage{
summaryGlmerSRMod(x)
}
\arguments{
\item{x}{a list}
}
\value{
An object of class \code{summaryGlmerSRMod}
}
\description{
Construct a summaryGlmerSRMod object
}

