% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmFit.R
\name{summaryGlmmFit}
\alias{summaryGlmmFit}
\title{Construct a summaryGlmmFit object}
\usage{
summaryGlmmFit(x)
}
\arguments{
\item{x}{a list}
}
\value{
An object of class \code{summaryGlmmFit}
}
\description{
Construct a summaryGlmmFit object
}
\keyword{internal}
