% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmMod.R
\name{summaryGlmmMod}
\alias{summaryGlmmMod}
\title{Construct a summaryGlmmMod object}
\usage{
summaryGlmmMod(x)
}
\arguments{
\item{x}{a list}
}
\value{
An object of class \code{summaryGlmmMod}
}
\description{
Construct a summaryGlmmMod object
}

