% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmsr.R
\docType{data}
\name{two_level}
\alias{two_level}
\title{A dataset simulated from a two-level model}
\format{An object of class \code{list} of length 3.}
\usage{
two_level
}
\description{
A dataset simulated from a two-level model
}
\keyword{datasets}

