% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmFit.R
\name{vcov.glmmFit}
\alias{vcov.glmmFit}
\title{Calculate Variance-Covariance Matrix for a glmmFit object}
\usage{
\method{vcov}{glmmFit}(object, ...)
}
\arguments{
\item{object}{glmmFit object}

\item{...}{ignored}
}
\value{
A matrix of the estimated covariances between the parameter estimates
}
\description{
Calculate Variance-Covariance Matrix for a glmmFit object
}
\keyword{internal}
