% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{find_lfun_glmm}
\alias{find_lfun_glmm}
\title{Find the log-likelihood function}
\usage{
find_lfun_glmm(modfr, method, control)
}
\arguments{
\item{modfr}{a model frame, the output of \code{find_modfr_glmm}}

\item{method}{the method used to approximate the likelihood. The options
are \code{"Laplace"}, \code{"AGQ"} (the adaptive Gaussian quadrature approximation,
from \code{lme4}), and \code{"SR"} (the sequential reduction approximation).}

\item{control}{a list of extra parameters controlling the approximation
to the likelihood. See 'Details' for more information.}
}
\description{
Find the log-likelihood function
}

