% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{find_lfun_glmm}
\alias{find_lfun_glmm}
\title{Find the log-likelihood function}
\usage{
find_lfun_glmm(modfr, method, control = NULL,
  lme4_control = set_lme4_control())
}
\arguments{
\item{modfr}{a model frame, the output of \code{find_modfr_glmm}}

\item{method}{the method used to approximate the likelihood. The options
are \code{"Laplace"}, \code{"AGQ"} (the adaptive Gaussian quadrature approximation,
from \code{lme4}), \code{"SR"} (the sequential reduction approximation)
and \code{"IS"} (an importance sampling approximation).}

\item{control}{a list of extra parameters controlling the approximation
to the likelihood. See 'Details' for more information.}

\item{lme4_control}{the result of a call to \code{lme4_control}, containing
control parameters passed to \code{lme4}. See \code{?lme4_control}.}
}
\description{
Find the log-likelihood function
}
