% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/subformula.R
\name{glFormulaSub}
\alias{glFormulaSub}
\title{Parse a formula (and possibly subformulas)}
\usage{
glFormulaSub(formula, data = NULL, family = gaussian, subforms = NULL)
}
\arguments{
\item{formula}{a two-sided linear formula object describing both the
    fixed-effects and fixed-effects part of the model, with the response
    on the left of a \code{~} operator and the terms, separated by
    \code{+} operators, on the right.  Random-effects terms are
    distinguished by vertical bars (\code{"|"}) separating expressions
    for design matrices from grouping factors.}

\item{data}{an optional data frame or environment containing the variables
named in \code{formula}, and in any of the subformulas contained in
subforms.}

\item{family}{a GLM family, see \code{\link[stats]{glm}} and
    \code{\link[stats]{family}}.}

\item{subforms}{an (optional) list of subformulas}
}
\description{
Parse a formula (and possibly subformulas)
}

