% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit.R
\name{glmerSR}
\alias{glmerSR}
\title{Fit a GLMM}
\usage{
glmerSR(formula, subformula = NULL, data = NULL, family = gaussian,
  control = glmmsrControl(), verbose = 0L, nAGQ = 1L, k = 0L,
  offset = NULL, devFunOnly = FALSE)
}
\arguments{
\item{formula}{a two-sided linear formula object describing both the
    fixed-effects and fixed-effects part of the model, with the response
    on the left of a \code{~} operator and the terms, separated by
    \code{+} operators, on the right.  Random-effects terms are
    distinguished by vertical bars (\code{"|"}) separating expressions
    for design matrices from grouping factors.}

\item{subformula}{a subformula, describing how a substituted variable
depends on covariates, or a list of subformulas, if there is more
than one \code{Sub()} term in \code{formula}.}

\item{data}{an optional data frame, list or environment containing the
variables named in \code{formula}, and in any of the subformulas.}

\item{family}{a GLM family, see \code{\link[stats]{glm}} and
    \code{\link[stats]{family}}.}

\item{control}{the output of a call to \code{\link{glmmsrControl}}, a list
containing control paramaters.}

\item{verbose}{integer scalar.  If \code{> 0} verbose output is
    generated during the optimization of the parameter estimates.  If
    \code{> 1} verbose output is generated during the individual PIRLS
    steps.}

\item{nAGQ}{integer scalar - the number of points per axis for
    evaluating the adaptive Gauss-Hermite approximation to the
    log-likelihood.  Defaults to 1, corresponding to the Laplace
    approximation.  Values greater than 1 produce greater accuracy in
    the evaluation of the log-likelihood at the expense of speed.  A
    value of zero uses a faster but less exact form of parameter
    estimation for GLMMs by optimizing the random effects and the
    fixed-effects coefficients in the penalized iteratively reweighted
    least squares step.}

\item{k}{integer scalar - the level of approximation used in the sequential
reduction approximation to the likelihood.}

\item{offset}{this can be used to specify an \emph{a priori} known
    component to be included in the linear predictor during
    fitting. This should be \code{NULL} or a numeric vector of length
    equal to the number of cases.  One or more \code{\link{offset}}
    terms can be included in the formula instead or as well, and if more
    than one is specified their sum is used.  See \code{\link{model.offset}}.}

\item{devFunOnly}{logical - return only the deviance evaluation
    function. Note that because the deviance function operates on
    variables stored in its environment, it may not return
    \emph{exactly} the same values on subsequent calls (but the results
    should always be within machine tolerance).}
}
\description{
Fit a GLMM
}

