% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_devfun.R
\name{mkGlmerDevfunSR}
\alias{mkGlmerDevfunSR}
\title{Make the deviance function}
\usage{
mkGlmerDevfunSR(fr, X, reTrms, family, nAGQ = 1L, k = 0L, verbose = 0L,
  control = glmmsrControl(), ...)
}
\arguments{
\item{fr}{A model frame containing the variables needed to create an
    \code{\link{lmerResp}} or \code{\link{glmResp}} instance.}

\item{X}{fixed-effects design matrix}

\item{reTrms}{information on random effects structure (see
    \code{\link{mkReTrms}}).}

\item{family}{a GLM family; see \code{\link[stats]{glm}}
    and \code{\link[stats]{family}}.}

\item{nAGQ}{number of Gauss-Hermite quadrature points}

\item{k}{integer scalar - the level of approximation used in the sequential
reduction approximation to the likelihood.}

\item{verbose}{print output?}

\item{control}{a list giving
    \describe{
      \item{for \code{[g]lFormula}:}{all
	options for running the model, see \code{\link{lmerControl}};}
      \item{for \code{mkLmerDevfun,mkGlmerDevfun}:}{options
	for the inner optimization step;}
      \item{for \code{optimizeLmer} and \code{optimizeGlmer}:}{control
	parameters for nonlinear optimizer (typically inherited from the
	\dots argument to \code{\link{lmerControl}}).}
	% FIXME: reference optCtrl
    }
  }

\item{...}{other potential arguments; for \code{optimizeLmer} and
    \code{optimizeGlmer}, these are passed to internal function
    \code{optwrap}, which has relevant parameters \code{calc.derivs}
    and \code{use.last.params} (see \code{\link{lmerControl}}).}
}
\description{
Make the deviance function
}

