% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{optimizeGlmm}
\alias{optimizeGlmm}
\title{Maximize the approximated log-likelihood}
\usage{
optimizeGlmm(lfun, p_beta, p_theta, init_na = NULL, verbose = 1L)
}
\arguments{
\item{lfun}{the approximated loglikelihood function}

\item{p_beta}{the number of covariates}

\item{p_theta}{the number of random effects}

\item{init_na}{a normal approximation to the log-likelihood surface}

\item{verbose}{controls how much detail to print out while fitting the model.
For verbose = 0, print nothing. For verbose = 1 (the default), print
output approximately once a second during model fitting.}
}
\value{
A list, containing the parameter estimate and variance matrix
}
\description{
Find the approximate maximum likelihood estimate, and finds
a normal approximation to the likelihood surface
}

