% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmmsr.R
\docType{data}
\name{three_level}
\alias{three_level}
\title{A simulated three-level dataset}
\format{\preformatted{List of 4
 $ response : int [1:200] 0 1 0 0 0 1 1 0 0 1 ...
 $ covariate: int [1:200] 0 0 1 1 0 1 1 1 1 0 ...
 $ cluster  : int [1:200] 1 1 2 2 3 3 4 4 5 5 ...
 $ group    : int [1:200] 1 1 1 1 2 2 2 2 3 3 ...
}}
\usage{
three_level
}
\description{
A simulated three-level dataset
}
\keyword{datasets}

