% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmmsr.R
\docType{data}
\name{two_level}
\alias{two_level}
\title{A simulated two-level dataset}
\format{\preformatted{List of 3
 $ response : int [1:100] 1 1 0 0 1 0 0 0 1 1 ...
 $ covariate: int [1:100] 0 0 1 1 0 1 1 1 1 0 ...
 $ cluster  : int [1:100] 1 1 2 2 3 3 4 4 5 5 ...
}}
\usage{
two_level
}
\description{
A simulated two-level dataset
}
\keyword{datasets}

